Move employee types to seperate file and go to 404 page if employee is not found
