+++ app/components/employeeCV.tsx
@@ -1,8 +1,8 @@
 "use client";
 import React, { useEffect, useState } from "react";
-import { EmployeeCV } from "../function/Employees";
 import { ErrorBoundary } from "./error";
 import dynamic from "next/dynamic";
+import Error from "next/error";
 import { Loading } from "./home";
 import { useChatStore } from "../store";
 import { isMobileScreen } from "../utils";
@@ -10,6 +10,7 @@ import styles from "../components/home.module.scss";
 import specificStyles from "../components/employeeCv.module.scss";
 import Sidebar from "./sidebar";
 import CV from "./CV";
+import { EmployeeCVDetails } from "../salesGPT/types";
 
 const Settings = dynamic(async () => (await import("./settings")).Settings, {
   loading: () => <Loading noLogo />,
@@ -26,10 +27,10 @@ const useHasHydrated = () => {
 };
 
 type SummaryOfQualificationProps = {
-  employeeAlias: string;
+  employee: EmployeeCVDetails | undefined;
 };
 
-function _EmployeeCV({ employeeAlias }: SummaryOfQualificationProps) {
+function _EmployeeCV({ employee }: SummaryOfQualificationProps) {
   const config = useChatStore((state) => state.config);
 
   // Setting
@@ -42,10 +43,14 @@ function _EmployeeCV({ employeeAlias }: SummaryOfQualificationProps) {
   if (isLoading) {
     return <Loading />;
   }
+  if (!employee) {
+    return <Error statusCode={404} />;
+  }
 
   async function handleButtonClick(requirementText: string): Promise<void> {
     setIsLoading(true);
     const requirements = requirementText.split("\n").filter((s) => s.length);
+    const employeeAlias = employee?.email.split("@")[0];
     await fetch("/api/chewbacca/generateSummaryOfQualifications", {
       method: "POST",
       headers: {
@@ -85,6 +90,9 @@ function _EmployeeCV({ employeeAlias }: SummaryOfQualificationProps) {
           <Settings closeSettings={() => setOpenSettings(false)} />
         ) : (
           <>
+            <div>
+              <p>Konsulent: {employee.email.split("@")[0]}</p>
+            </div>
             <textarea
               className={specificStyles.RequirementInput}
               placeholder={
@@ -93,7 +101,6 @@ function _EmployeeCV({ employeeAlias }: SummaryOfQualificationProps) {
               value={requirementText}
               onChange={(event) => setRequirementText(event.target.value)}
             ></textarea>
-            <CV GPTResponse={employeeAlias} />
             <button onClick={async () => handleButtonClick(requirementText)}>
               Generer oppsummering av kvalifikasjoner
             </button>
@@ -105,14 +112,12 @@ function _EmployeeCV({ employeeAlias }: SummaryOfQualificationProps) {
   );
 }
 
-export default function EmployeeCV({
-  employeeAlias,
-}: SummaryOfQualificationProps) {
+export default function EmployeeCV({ employee }: SummaryOfQualificationProps) {
   return (
     <ErrorBoundary
       fallback={<p> Something went wrong with the EmployeeCV! </p>}
     >
-      <_EmployeeCV employeeAlias={employeeAlias} />
+      <_EmployeeCV employee={employee} />
     </ErrorBoundary>
   );
 }
