+++ app/salesGPT/[employeeAlias]/page.tsx
@@ -1,10 +1,12 @@
 import { redirect } from "next/navigation";
+import Error from "next/error";
 
 import { getServerSession } from "next-auth/next";
 
 import { authOptions } from "../../api/auth/auth-options";
 import EmployeeCV from "@/app/components/employeeCV";
 import { CustomSession } from "@/app/api/auth/[...nextauth]/typing";
+import { getEmployeeCVData } from "@/app/function/Employees";
 
 export default async function App({
   params,
@@ -15,5 +17,10 @@ export default async function App({
   if (!session) {
     return redirect("/api/auth/signin");
   }
-  return <EmployeeCV employeeAlias={params.employeeAlias} />;
+  const employee = await getEmployeeCVData(
+    params.employeeAlias,
+    session.access_token,
+  );
+
+  return <EmployeeCV employee={employee} />;
 }
